tests: add endpoint_name to FlowCreate in simple API test and add tests using `endpoint_name`
